#7 - Added Delete action for Alert Requests including tests

+++ Metal Mate MVC/Metal Mate MVC/Services/AlertRequestService.cs
@@ -10,7 +10,7 @@ namespace Metal_Mate_MVC.Services
         Task AddAsync(AlertRequest alertRequest);
         Task SaveAsync(AlertRequest alertRequest);
         Task<AlertRequest?> GetByIdAsync(int id);
-        Task DeleteAsync(AlertRequest alertRequest);
+        Task<bool> DeleteAsync(int id);
     }
 
     public class AlertRequestService : IAlertRequestService
@@ -47,10 +47,19 @@ namespace Metal_Mate_MVC.Services
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(AlertRequest alertRequest)
+        public async Task<bool> DeleteAsync(int id)
         {
+            var alertRequest = await _context.AlertRequests.FindAsync(id);
+
+            if (alertRequest == null)
+            {
+                return false;
+            }
+
             _context.AlertRequests.Remove(alertRequest);
             await _context.SaveChangesAsync();
+
+            return true;
         }
 
     }
